Extract parameter actions
