+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -1,4 +1,5 @@
 export * from "./actions";
 export * from "./core";
+export * from "./parameters";
 export * from "./ui";
 export * from "./writeback";
